Adapter, Magpie, all chains, wip interactions
