+++ src/adapters/magpie/common/stake.ts
