+++ src/adapters/magpie/index.ts
@@ -1,12 +1,16 @@
 import type { Adapter } from '@lib/adapter'
 
 import * as arbitrum from './arbitrum'
-import * as bsc from './bsc'
+// import * as bsc from './bsc'
+// import * as ethereum from './ethereum'
+// import * as optimism from './optimism'
 
 const adapter: Adapter = {
   id: 'magpie',
   arbitrum,
-  bsc,
+  // bsc,
+  // ethereum,
+  // optimism,
 }
 
 export default adapter
